se separa el código para crear un nuevo hook useProfileConfigCatchment
