+++ src/hooks/useCatchmentPoint.ts
@@ -18,30 +18,10 @@ export type CatchmentPoint = {
 
 type NewCatchmentPoint = Omit<CatchmentPoint, 'id' | 'created' | 'modified'>;
 
-export type ProfileDataConfig = {
-  id: number;
-  created: string;
-  modified: string;
-  token_service: string;
-  d1: string;
-  d2: string;
-  d3: string;
-  d4: string;
-  d5: string;
-  d6: number;
-  is_telemetry: boolean;
-  date_start_telemetry: string | null;
-  date_delivery_act: string | null;
-  point_catchment: number;
-};
-
-type NewProfileDataConfig = Omit<ProfileDataConfig, 'id' | 'created' | 'modified'>;
-
 export const useCatchmentPoint = () => {
   const { loading, error, fetchData } = useApi();
   const [catchmentPoints, setCatchmentPoints] = useState<CatchmentPoint[]>([]);
   const [currentPoint, setCurrentPoint] = useState<CatchmentPoint | null>(null);
-  const [profileDataConfigs, setProfileDataConfigs] = useState<ProfileDataConfig[]>([]);
 
     const getAll = async () => {
     const data = await fetchData<{
@@ -108,29 +88,6 @@ export const useCatchmentPoint = () => {
     }
   };
 
-  const getProfileDataConfigs = async () => {
-    const data = await fetchData<{
-      count: number;
-      next: string | null;
-      previous: string | null;
-      results: ProfileDataConfig[];
-    }>('/api/profile_data_config_catchment/');
-
-    if (data) setProfileDataConfigs(data.results);
-    return data;
-  };
-
-  const createProfileDataConfig = async (data: NewProfileDataConfig) => {
-    try {
-      const response = await axios.post<ProfileDataConfig>('/api/profile_data_config_catchment/', data);
-      setProfileDataConfigs(prev => [...prev, response.data]);
-      return response.data;
-    } catch (err) {
-      console.error(err);
-      return null;
-    }
-  };
-
   return {
     loading,
     error,
@@ -142,8 +99,5 @@ export const useCatchmentPoint = () => {
     update,
     patch,
     remove,
-    profileDataConfigs,
-    getProfileDataConfigs,
-    createProfileDataConfig,
   };
 };
